requirements.txt
fiona>=1.9.1
gdal>=3.6.2
geopandas>=0.12.2
geopy>=2.3.0
gpytorch>=1.9.1
jupyter>=1.0.0
matplotlib>=3.7.1
matplotlib-inline>=0.1.6
numpy>=1.23.5
pandas>=1.5.3
python>=3.9.16
pytorch>=1.12.1
scikit-mobility>=1.3.1
scikit-learn>=1.2.2
shapely>=2.0.1
torchaudio>=0.12.1
torchvision>=0.13.1
tqdm>=4.65.0
notebook>=6.5.3
holidays>=0.21
